refactor: separate the source code of success and error  modal into a module modal.js
