+++ js/form.js
@@ -1,5 +1,6 @@
 import { sendData } from './network.js';
 import { classEffectToIdEffect, idEffectToClassEffect } from './util.js';
+import { openSuccessModal, openErrorModal } from './modal.js';
 
 const DEFAULT_SCALE_PHOTO = '100%';
 
@@ -23,8 +24,6 @@ const sliderFieldset = document.querySelector('.img-upload__effect-level');
 const imgUploadForm = document.querySelector('#upload-select-image');
 const imgUploadSubmit = imgUploadForm.querySelector('#upload-submit');
 
-const successModalTemplate = document.querySelector('#success').content.querySelector('.success');
-const errorModalTemplate = document.querySelector('#error').content.querySelector('.error');
 const effectNoneRadioInput = effectsList.querySelector('#effect-none');
 
 const ScaleParameter = {
@@ -132,73 +131,16 @@ const onEffectsListChange = (evt) => {
   }
 };
 
-const onSuccessModalEscKeydown = (evt) => {
-  if (evt.key === 'Escape') {
-    onSuccessModalButtonClick();
-  }
-};
-
-const onOutsideSuccessModalClick = (evt) => {
-  const successInnerModal = document.querySelector('.success__inner');
-  if (!(evt.target === successInnerModal) && !successInnerModal.contains(evt.target)) {
-    onSuccessModalButtonClick();
-  }
-};
-
-function onSuccessModalButtonClick() {
-  const successModal = document.querySelector('.success');
-  const successModalButton = successModal.querySelector('.success__button');
-  successModalButton.removeEventListener('click', onSuccessModalButtonClick);
-  successModal.remove();
-  document.removeEventListener('keydown', onSuccessModalEscKeydown);
-  document.removeEventListener('click', onOutsideSuccessModalClick);
-  document.addEventListener('keydown', onModalEscKeydown);
-  onCancelFormButtonClick();
-}
-
-const onErrorModalEscKeydown = (evt) => {
-  if (evt.key === 'Escape') {
-    onErrorModalButtonClick();
-  }
-};
-
-const onOutsideErrorModalClick = (evt) => {
-  const errorInnerModal = document.querySelector('.error__inner');
-  if (!(evt.target === errorInnerModal) && !errorInnerModal.contains(evt.target)) {
-    onErrorModalButtonClick();
-  }
-};
-
-function onErrorModalButtonClick() {
-  const errorModal = document.querySelector('.error');
-  const errorModalButton = errorModal.querySelector('.error__button');
-  errorModalButton.removeEventListener('click', onErrorModalButtonClick);
-  errorModal.remove();
-  document.removeEventListener('keydown', onErrorModalEscKeydown);
-  document.removeEventListener('click', onOutsideErrorModalClick);
-  document.addEventListener('keydown', onModalEscKeydown);
-}
-
 const onSuccessSendData = () => {
   imgUploadSubmit.removeAttribute('disabled', '');
-  const successModal = successModalTemplate.cloneNode(true);
-  const successModalButton = successModal.querySelector('.success__button');
-  successModalButton.addEventListener('click', onSuccessModalButtonClick);
-  document.addEventListener('keydown', onSuccessModalEscKeydown);
-  document.addEventListener('click', onOutsideSuccessModalClick);
   document.removeEventListener('keydown', onModalEscKeydown);
-  body.append(successModal);
+  openSuccessModal();
 };
 
 const onErrorSendData = () => {
   imgUploadSubmit.removeAttribute('disabled', '');
-  const errorModal = errorModalTemplate.cloneNode(true);
-  const errorModalButton = errorModal.querySelector('.error__button');
-  errorModalButton.addEventListener('click', onErrorModalButtonClick);
-  document.addEventListener('keydown', onErrorModalEscKeydown);
-  document.addEventListener('click', onOutsideErrorModalClick);
   document.removeEventListener('keydown', onModalEscKeydown);
-  body.append(errorModal);
+  openErrorModal();
 };
 
 const onFormSubmit = (evt) => {
@@ -259,4 +201,4 @@ const createEventFormHandlers = () => {
   photoUploadButton.addEventListener('change', onUploadPhotoButtonChange);
 };
 
-export {createEventFormHandlers};
+export {createEventFormHandlers, onCancelFormButtonClick, onModalEscKeydown};
